revert to original, unformatted, ogb ppa code

+++ tst/ogb/original/ppa/conv.py
@@ -1,38 +1,36 @@
 import torch
 from torch_geometric.nn import MessagePassing
 import torch.nn.functional as F
-from torch_geometric.nn import global_add_pool
+from torch_geometric.nn import global_mean_pool, global_add_pool
 from torch_geometric.utils import degree
 
+import math
 
-
-# GIN convolution along the graph structure
+### GIN convolution along the graph structure
 class GINConv(MessagePassing):
     def __init__(self, emb_dim):
         '''
             emb_dim (int): node embedding dimensionality
         '''
 
-        super(GINConv, self).__init__(aggr="add")
+        super(GINConv, self).__init__(aggr = "add")
 
-        self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
-                                       torch.nn.ReLU(), torch.nn.Linear(2 * emb_dim, emb_dim))
+        self.mlp = torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), torch.nn.Linear(2*emb_dim, emb_dim))
         self.eps = torch.nn.Parameter(torch.Tensor([0]))
-
+        
         self.edge_encoder = torch.nn.Linear(7, emb_dim)
 
     def forward(self, x, edge_index, edge_attr):
-        edge_embedding = self.edge_encoder(edge_attr)
-        out = self.mlp((1 + self.eps) * x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
-
+        edge_embedding = self.edge_encoder(edge_attr)    
+        out = self.mlp((1 + self.eps) *x + self.propagate(edge_index, x=x, edge_attr=edge_embedding))
+    
         return out
 
     def message(self, x_j, edge_attr):
         return F.relu(x_j + edge_attr)
-
+        
     def update(self, aggr_out):
         return aggr_out
-
 
 ### GCN convolution along the graph structure
 class GCNConv(MessagePassing):
@@ -49,15 +47,14 @@ class GCNConv(MessagePassing):
 
         row, col = edge_index
 
-        # edge_weight = torch.ones((edge_index.size(1), ), device=edge_index.device)
-        deg = degree(row, x.size(0), dtype=x.dtype) + 1
+        #edge_weight = torch.ones((edge_index.size(1), ), device=edge_index.device)
+        deg = degree(row, x.size(0), dtype = x.dtype) + 1
         deg_inv_sqrt = deg.pow(-0.5)
         deg_inv_sqrt[deg_inv_sqrt == float('inf')] = 0
 
         norm = deg_inv_sqrt[row] * deg_inv_sqrt[col]
 
-        return self.propagate(edge_index, x=x, edge_attr=edge_embedding, norm=norm) + F.relu(
-            x + self.root_emb.weight) * 1. / deg.view(-1, 1)
+        return self.propagate(edge_index, x=x, edge_attr = edge_embedding, norm=norm) + F.relu(x + self.root_emb.weight) * 1./deg.view(-1,1)
 
     def message(self, x_j, edge_attr, norm):
         return norm.view(-1, 1) * F.relu(x_j + edge_attr)
@@ -72,8 +69,7 @@ class GNN_node(torch.nn.Module):
     Output:
         node representations
     """
-
-    def __init__(self, num_layer, emb_dim, drop_ratio=0.5, JK="last", residual=False, gnn_type='gin'):
+    def __init__(self, num_layer, emb_dim, drop_ratio = 0.5, JK = "last", residual = False, gnn_type = 'gin'):
         '''
             emb_dim (int): node embedding dimensionality
             num_layer (int): number of GNN message passing layers
@@ -90,7 +86,7 @@ class GNN_node(torch.nn.Module):
         if self.num_layer < 2:
             raise ValueError("Number of GNN layers must be greater than 1.")
 
-        self.node_encoder = torch.nn.Embedding(1, emb_dim)  # uniform input node embedding
+        self.node_encoder = torch.nn.Embedding(1, emb_dim) # uniform input node embedding
 
         ###List of GNNs
         self.convs = torch.nn.ModuleList()
@@ -103,11 +99,12 @@ class GNN_node(torch.nn.Module):
                 self.convs.append(GCNConv(emb_dim))
             else:
                 ValueError('Undefined GNN type called {}'.format(gnn_type))
-
+                
             self.batch_norms.append(torch.nn.BatchNorm1d(emb_dim))
 
     def forward(self, batched_data):
         x, edge_index, edge_attr, batch = batched_data.x, batched_data.edge_index, batched_data.edge_attr, batched_data.batch
+
 
         ### computing input node embedding
 
@@ -118,10 +115,10 @@ class GNN_node(torch.nn.Module):
             h = self.batch_norms[layer](h)
 
             if layer == self.num_layer - 1:
-                # remove relu for the last layer
-                h = F.dropout(h, self.drop_ratio, training=self.training)
+                #remove relu for the last layer
+                h = F.dropout(h, self.drop_ratio, training = self.training)
             else:
-                h = F.dropout(F.relu(h), self.drop_ratio, training=self.training)
+                h = F.dropout(F.relu(h), self.drop_ratio, training = self.training)
 
             if self.residual:
                 h += h_list[layer]
@@ -145,8 +142,7 @@ class GNN_node_Virtualnode(torch.nn.Module):
     Output:
         node representations
     """
-
-    def __init__(self, num_layer, emb_dim, drop_ratio=0.5, JK="last", residual=False, gnn_type='gin'):
+    def __init__(self, num_layer, emb_dim, drop_ratio = 0.5, JK = "last", residual = False, gnn_type = 'gin'):
         '''
             emb_dim (int): node embedding dimensionality
         '''
@@ -161,7 +157,7 @@ class GNN_node_Virtualnode(torch.nn.Module):
         if self.num_layer < 2:
             raise ValueError("Number of GNN layers must be greater than 1.")
 
-        self.node_encoder = torch.nn.Embedding(1, emb_dim)  # uniform input node embedding
+        self.node_encoder = torch.nn.Embedding(1, emb_dim) # uniform input node embedding
 
         ### set the initial virtual node embedding to 0.
         self.virtualnode_embedding = torch.nn.Embedding(1, emb_dim)
@@ -186,19 +182,16 @@ class GNN_node_Virtualnode(torch.nn.Module):
             self.batch_norms.append(torch.nn.BatchNorm1d(emb_dim))
 
         for layer in range(num_layer - 1):
-            self.mlp_virtualnode_list.append(
-                torch.nn.Sequential(torch.nn.Linear(emb_dim, 2 * emb_dim), torch.nn.BatchNorm1d(2 * emb_dim),
-                                    torch.nn.ReLU(), \
-                                    torch.nn.Linear(2 * emb_dim, emb_dim), torch.nn.BatchNorm1d(emb_dim),
-                                    torch.nn.ReLU()))
+            self.mlp_virtualnode_list.append(torch.nn.Sequential(torch.nn.Linear(emb_dim, 2*emb_dim), torch.nn.BatchNorm1d(2*emb_dim), torch.nn.ReLU(), \
+                                                    torch.nn.Linear(2*emb_dim, emb_dim), torch.nn.BatchNorm1d(emb_dim), torch.nn.ReLU()))
+
 
     def forward(self, batched_data):
 
         x, edge_index, edge_attr, batch = batched_data.x, batched_data.edge_index, batched_data.edge_attr, batched_data.batch
 
         ### virtual node embeddings for graphs
-        virtualnode_embedding = self.virtualnode_embedding(
-            torch.zeros(batch[-1].item() + 1).to(edge_index.dtype).to(edge_index.device))
+        virtualnode_embedding = self.virtualnode_embedding(torch.zeros(batch[-1].item() + 1).to(edge_index.dtype).to(edge_index.device))
 
         h_list = [self.node_encoder(x)]
         for layer in range(self.num_layer):
@@ -210,10 +203,10 @@ class GNN_node_Virtualnode(torch.nn.Module):
 
             h = self.batch_norms[layer](h)
             if layer == self.num_layer - 1:
-                # remove relu for the last layer
-                h = F.dropout(h, self.drop_ratio, training=self.training)
+                #remove relu for the last layer
+                h = F.dropout(h, self.drop_ratio, training = self.training)
             else:
-                h = F.dropout(F.relu(h), self.drop_ratio, training=self.training)
+                h = F.dropout(F.relu(h), self.drop_ratio, training = self.training)
 
             if self.residual:
                 h = h + h_list[layer]
@@ -227,12 +220,9 @@ class GNN_node_Virtualnode(torch.nn.Module):
                 ### transform virtual nodes using MLP
 
                 if self.residual:
-                    virtualnode_embedding = virtualnode_embedding + F.dropout(
-                        self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.drop_ratio,
-                        training=self.training)
+                    virtualnode_embedding = virtualnode_embedding + F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.drop_ratio, training = self.training)
                 else:
-                    virtualnode_embedding = F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp),
-                                                      self.drop_ratio, training=self.training)
+                    virtualnode_embedding = F.dropout(self.mlp_virtualnode_list[layer](virtualnode_embedding_temp), self.drop_ratio, training = self.training)
 
         ### Different implementations of Jk-concat
         if self.JK == "last":
@@ -241,7 +231,7 @@ class GNN_node_Virtualnode(torch.nn.Module):
             node_representation = 0
             for layer in range(self.num_layer):
                 node_representation += h_list[layer]
-
+        
         return node_representation
